Add files via upload
The name of the robot was inspired by the orange colour. The Robot was made with a special material called Alcobot which is a infusion of plastic and aluminium giving it it's rigidity and light weight nature. This was a Bluetooth controlled robot. Initially it was build for a competition held by the Robotic Club of BRAC University. I made it up to the semi-final round. In the future, I plan to infuse a camera and control it using Bluetooth to deliver little things from one room to another and carry out my house-hold chores.